fixed esc functionality, fixed modal functionality
